+++ src/Benzene.Extras/Broadcast/IBroadcastEventChecker.cs
@@ -1,8 +1,0 @@
-﻿using Benzene.Abstractions.Messages;
-
-namespace Benzene.Extras.Broadcast;
-
-public interface IBroadcastEventChecker : IMessageDefinitionFinder<IMessageDefinition>
-{
-    bool Check<T>(string topic, T payload);
-}