+++ src/Benzene.Extras/Broadcast/IEventSender.cs
@@ -1,6 +1,0 @@
-﻿namespace Benzene.Extras.Broadcast;
-
-public interface IEventSender
-{
-    Task SendAsync<T>(string topic, T payload);
-}